libathemecore/linker.c: use correct data type for buffer sizes

+++ include/atheme/linker.h
@@ -12,6 +12,6 @@
 
 #include <atheme/stdheaders.h>
 
-mowgli_module_t *linker_open_ext(const char *path, char *errbuf, int errlen);
+mowgli_module_t *linker_open_ext(const char *path, char *errbuf, size_t errlen);
 
 #endif /* !ATHEME_INC_LINKER_H */
